refactor: move mapper in interface
