+++ front/src/pages/people/People.tsx
@@ -4,9 +4,12 @@ import Table from '../../components/table/Table';
 import styled from '@emotion/styled';
 import { peopleColumns } from './people-table';
 import { gql, useQuery } from '@apollo/client';
-import { GraphqlPerson, Person } from '../../interfaces/person.interface';
+import {
+  GraphqlPerson,
+  Person,
+  mapPerson,
+} from '../../interfaces/person.interface';
 import { defaultData } from './default-data';
-import { mapPerson } from './mapper';
 
 const StyledPeopleContainer = styled.div`
   display: flex;
